Moved mapping builders into separate module
