+++ src/mapping/mod.rs
@@ -1,3 +1,4 @@
+pub mod build;
 pub mod parse;
 
 use std::collections::HashMap;
